Added collaborators

+++ app/controllers/wikis_controller.rb
@@ -1,30 +1,27 @@
-include ApplicationHelper
-
 class WikisController < ApplicationController
   before_action :authenticate_user!
   
   def index
-    @wikis = Wiki.all
+    authorize(Wiki)
+    @wikis = policy_scope(Wiki)
   end
 
   def show
     @wiki = Wiki.find(params[:id])
-    
-    if @wiki.private && current_user.standard?
-      flash[:alert] = "You must be a priemium user to view this Wiki."
-      redirect_to root_path
-    end
+    authorize @wiki
   end
 
   def new
     @user = current_user
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @user = current_user
     
     @wiki = Wiki.new(user: current_user)
+    authorize @wiki
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
     @wiki.private = params[:wiki][:private]
@@ -36,34 +33,29 @@ class WikisController < ApplicationController
       flash.now[:alert] = "There was an error saving the wiki. Please try again."
       render :new
     end
-     
-    # @wiki = current_user.wikis.create(wiki_params)
-    # authorize @wiki
-    # # adding owner to collaborator list
-    # # collaborator = @wiki.collaborators.build(user_id: current_user.id)
   end
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
-    @wiki.private = params[:wiki][:private]
+    @wiki.assign_attributes(wiki_params)
 
     if @wiki.save
-      flash[:notice] = "Wiki was updated successfully."
+      flash[:notice] = "You wiki has been updated."
       redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+      flash.now[:alert] = "There was an Error editing your Wiki."
       render :edit
     end
   end
   
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
@@ -77,16 +69,6 @@ class WikisController < ApplicationController
   private
   
   def wiki_params
-    params.require(:wiki).permit(:title, :body, :private, :user)
+    params.require(:wiki).permit(:title, :body, :private, :user_id, :user_ids => [])
   end
-  
-  def authorize_user
-    @user = User.find(params[:id])
-    authorize @user
-    unless current_user?
-      flash[:alert] = "You must be a user to do that."
-      redirect_to wikis_path
-    end
-  end
-
 end
